+++ src/optimum/rbln/transformers/models/siglip/modeling_siglip.py
@@ -147,7 +147,7 @@ class RBLNSiglipVisionModel(RBLNModel):
                 f"Please compile again with the correct argument."
             )
 
-        output = super().forward(pixel_values, return_dict=return_dict, **kwargs)
+        output = super().forward(pixel_values.to(self.rbln_config.dtype), return_dict=return_dict, **kwargs)
         return output
 
     def _prepare_output(self, output, return_dict):
